+++ ui/src/app/modules/status/status.component.ts
@@ -111,7 +111,16 @@ export class StatusComponent implements OnInit, OnDestroy {
         if (!layout.length) {
           return this.resetLayout()
         }
-        this.setLayout(layout)
+        this.setLayout(layout.map((item: any) => {
+          // some name changes
+          if (item.component === 'ChildBridgeWidgetComponent') {
+            item.component = 'HomekitBridgesWidgetComponent'
+          }
+          if (item.component === 'HomebridgeStatusWidgetComponent') {
+            item.component = 'UpdateCentreWidgetComponent'
+          }
+          return item
+        }))
       },
     )
   }