+++ ui/src/app/modules/status/widget-visibility/widget-visibility.component.ts
@@ -24,8 +24,8 @@ export class WidgetVisibilityComponent implements OnInit {
   ngOnInit() {
     const allWidgets = [
       {
-        name: this.$translate.instant('status.widget.add.label_homebridge_status'),
-        component: 'HomebridgeStatusWidgetComponent',
+        name: this.$translate.instant('status.services.updates'),
+        component: 'UpdateCentreWidgetComponent',
         hidden: false,
         cols: 10,
         rows: 3,
@@ -50,8 +50,8 @@ export class WidgetVisibilityComponent implements OnInit {
         hideOnMobile: false,
       },
       {
-        name: 'Child Bridge Status',
-        component: 'ChildBridgeWidgetComponent',
+        name: this.$translate.instant('child_bridge.bridges'),
+        component: 'HomekitBridgesWidgetComponent',
         hidden: !this.$settings.env.serviceMode,
         cols: 5,
         rows: 9,