+++ ui/src/app/modules/status/widgets/system-info-widget/system-info-widget.component.ts
@@ -68,30 +68,4 @@ export class SystemInfoWidgetComponent implements OnInit {
 
     ref.componentInstance.ctaButtonLink = 'https://github.com/homebridge/homebridge-config-ui-x/wiki/How-To-Swap-From-Standalone-Mode-to-Service-Mode'
   }
-
-  nodeUpdateModal() {
-    const ref = this.$modal.open(InformationComponent, {
-      size: 'lg',
-      backdrop: 'static',
-    })
-
-    ref.componentInstance.title = `${this.$translate.instant('status.widget.info.node_update_title')} - ${this.nodejsInfo.latestVersion}`
-    ref.componentInstance.message = this.$translate.instant('status.widget.info.node_update_message')
-    ref.componentInstance.ctaButtonLabel = this.$translate.instant('form.button_more_info')
-    ref.componentInstance.faIconClass = 'fab fa-fw fa-node-js primary-text'
-    ref.componentInstance.ctaButtonLink = 'https://github.com/homebridge/homebridge/wiki/How-To-Update-Node.js'
-  }
-
-  nodeUnsupportedModal() {
-    const ref = this.$modal.open(InformationComponent, {
-      size: 'lg',
-      backdrop: 'static',
-    })
-
-    ref.componentInstance.title = this.$translate.instant('status.widget.info.node_unsupp_title')
-    ref.componentInstance.message = this.$translate.instant('status.widget.info.node_unsupp_message')
-    ref.componentInstance.ctaButtonLabel = this.$translate.instant('form.button_more_info')
-    ref.componentInstance.faIconClass = 'fab fa-fw fa-node-js primary-text'
-    ref.componentInstance.ctaButtonLink = 'https://github.com/homebridge/homebridge/wiki/How-To-Update-Node.js'
-  }
 }